imx8mm_var_dart: spl: Change VDD_DRAM to 0.975v
According to latest datasheet IMX8MMCEC_Rev_0, the typical voltage
of VDD_DRAM for 1.5GHz DDR clock is 0.95v. Because BD71847MWV PMIC
does not support 0.95v output, we change the voltage to 0.975v as
the note in datasheet mentioned it is acceptable and supported.

See MLK-20902 commit from NXP.

Signed-off-by: Eran Matityahu <eran.m@variscite.com>