ddr: imx8m: Return error values from LPDDR4 training
In cases when the same SPL should run on boards with i.MX8M, that
differ in DDR configuration, it is necessary to try different
parameters and check if the training done by the firmware suceeds or
not.

Therefore we return the DDR training/initialization success to the
upper layer in order to be able to retry with different settings if
necessary.

This commit is ported from commit 83083febf55679ee0fc68ba55e9af43add277b58
by Frieder Schrempf <frieder.schrempf@kontron.de>.

Signed-off-by: Eran Matityahu <eran.m@variscite.com>