net: phy: mxl-8611x: fix ethernet-phy node handle
In newer versions of U-Boot, phy_get_ofnode(phydev) returns the
node of the phy, e.g. ethernet-phy@4.

In this U-Boot branch, phy_get_ofnode is returning a pointer to the
ethernet controller node.

To fix this, check if the node has a phy-handle property. If it does,
follow it to get the ethernet phy node.

Signed-off-by: Nate Drude <nate.d@variscite.com>